The HUD functions have gotten their own file because lots of code has been added.
Fields can now depend on a variable. The width drawn of an image can depend on your currenthealth, showing a progressbar.

Clicking events for objects (selects the object, displays a second healthbar if it has health) AND the HUD.
You can click on your own healthbar and you will get HP.