Moved FPS-measuring to the Canvas class, and improved it: FPS data are now calculated mean averages.
Improvements to the dirty rectangles code.

Found a bug (that still needs fixing): all objects on the current map are drawn, no matter where they are. That needs to be fixed.

Fixed the echo functions. innerHTML is not passed by reference. Stupid mistake

Started very early work of the autotile feature: added a new tileset and places a few tiles on the grassland map.